I found that the HAL implementation of a test for the CAN_FMR register is wrong.
In dm00031020, chapter 32.9.4, it is said that the register can take values up to 28 in order to have all filters assigned to can1.

However, in stm32F1XX_hal_can.h, the define IS_CAN_FILTER_BANK_DUAL tests it to be <=27.

I do believe this test is wrong and will fail a valid call to set filters if asserts are enabled.